Split ProductMedia into separate modules for better maintainability
Refactor the monolithic product-media.tsx and media-zoom.tsx into a
product-media/ directory with focused modules: index (orchestrator),
media-grid, media-slider, media-item, media-zoom, and utils.
