--- app/components/product/product-media.tsx
@@ -1,444 +1,0 @@
-import {
-  ArrowLeftIcon,
-  ArrowRightIcon,
-  CaretDownIcon,
-  CaretUpIcon,
-  VideoCameraIcon,
-} from "@phosphor-icons/react";
-import { cva, type VariantProps } from "class-variance-authority";
-import clsx from "clsx";
-import { useEffect, useState } from "react";
-import type {
-  Media_MediaImage_Fragment,
-  Media_Video_Fragment,
-  MediaFragment,
-  ProductQuery,
-  ProductVariantFragment,
-} from "storefront-api.generated";
-import { FreeMode, Navigation, Pagination, Thumbs } from "swiper/modules";
-import { Swiper, type SwiperClass, SwiperSlide } from "swiper/react";
-import { Image } from "~/components/image";
-import type { ImageAspectRatio } from "~/types/others";
-import { cn } from "~/utils/cn";
-import { calculateAspectRatio } from "~/utils/image";
-import { getVariantGroupedMedia } from "~/utils/variant-media";
-import { ZoomButton, ZoomModal } from "./media-zoom";
-
-const variants = cva(
-  [
-    "grid w-full justify-start gap-2 lg:gap-1",
-    "lg:grid-cols-1",
-    "grid-flow-col lg:grid-flow-row",
-    "scroll-px-6 overflow-x-scroll md:overflow-x-auto",
-    "snap-x snap-mandatory",
-  ],
-  {
-    variants: {
-      gridSize: {
-        "1x1": "",
-        "2x2": "2xl:grid-cols-2",
-        mix: "2xl:grid-cols-2",
-      },
-    },
-  },
-);
-
-export interface ProductMediaProps extends VariantProps<typeof variants> {
-  mediaLayout: "grid" | "slider";
-  imageAspectRatio?: ImageAspectRatio;
-  showThumbnails: boolean;
-  selectedVariant: ProductVariantFragment;
-  media: MediaFragment[];
-  enableZoom?: boolean;
-  zoomTrigger?: "image" | "button" | "both";
-  zoomButtonVisibility?: "always" | "hover";
-  groupMediaByVariant?: boolean;
-  groupByOption?: string;
-  product?: NonNullable<ProductQuery["product"]>;
-  initialMediaCount?: number;
-  showMoreText?: string;
-  showLessText?: string;
-}
-
-export function ProductMedia(props: ProductMediaProps) {
-  const {
-    mediaLayout: initialMediaLayout,
-    gridSize: initialGridSize,
-    showThumbnails,
-    imageAspectRatio,
-    selectedVariant,
-    media,
-    enableZoom,
-    zoomTrigger = "button",
-    zoomButtonVisibility = "hover",
-    groupMediaByVariant,
-    groupByOption,
-    product,
-    initialMediaCount = 0,
-    showMoreText = "Show more",
-    showLessText = "Show less",
-  } = props;
-
-  let displayMedia = media;
-  if (groupMediaByVariant && product && groupByOption) {
-    displayMedia = getVariantGroupedMedia({
-      allMedia: media,
-      selectedVariant,
-      product,
-      groupByOption,
-    });
-  }
-
-  const [swiper, setSwiper] = useState<SwiperClass | null>(null);
-  const [thumbsSwiper, setThumbsSwiper] = useState<SwiperClass | null>(null);
-  const [zoomMediaId, setZoomMediaId] = useState<string | null>(null);
-  const [zoomModalOpen, setZoomModalOpen] = useState(false);
-  const [expanded, setExpanded] = useState(false);
-
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation> --- IGNORE ---
-  useEffect(() => {
-    if (selectedVariant && swiper) {
-      const index = getSelectedVariantMediaIndex(displayMedia, selectedVariant);
-      if (index !== swiper.activeIndex) {
-        swiper.slideTo(index);
-      }
-    }
-  }, [selectedVariant]);
-
-  // biome-ignore lint/correctness/useExhaustiveDependencies: <explanation> --- IGNORE ---
-  useEffect(() => {
-    setExpanded(false);
-  }, [selectedVariant]);
-
-  let mediaLayout = initialMediaLayout;
-  let gridSize = initialGridSize;
-  if (displayMedia.length === 1) {
-    mediaLayout = "grid";
-    gridSize = "1x1";
-  }
-
-  const shouldShowButton =
-    enableZoom && (zoomTrigger === "button" || zoomTrigger === "both");
-  const canClickImage =
-    enableZoom && (zoomTrigger === "image" || zoomTrigger === "both");
-
-  if (mediaLayout === "grid") {
-    const shouldLimitMedia =
-      initialMediaCount > 0 && displayMedia.length > initialMediaCount;
-    const visibleMedia =
-      shouldLimitMedia && !expanded
-        ? displayMedia.slice(0, initialMediaCount)
-        : displayMedia;
-    const hiddenCount = displayMedia.length - visibleMedia.length;
-
-    return (
-      <>
-        <div className="relative">
-          <div className={variants({ gridSize })}>
-            {visibleMedia.map((med, idx) => {
-              const isLast = idx === visibleMedia.length - 1;
-              return (
-                <div
-                  key={med.id}
-                  className={clsx(
-                    "group relative",
-                    gridSize === "2x2" &&
-                      isLast &&
-                      visibleMedia.length % 2 === 1 &&
-                      "2xl:col-span-2",
-                    gridSize === "mix" &&
-                      (idx % 3 === 0 || (isLast && idx % 3 === 1)) &&
-                      "lg:col-span-2",
-                  )}
-                >
-                  <div
-                    onClick={
-                      canClickImage
-                        ? () => {
-                            setZoomMediaId(med.id);
-                            setZoomModalOpen(true);
-                          }
-                        : undefined
-                    }
-                    className={canClickImage ? "cursor-zoom-in" : ""}
-                  >
-                    <Media
-                      media={med}
-                      imageAspectRatio={imageAspectRatio}
-                      index={idx}
-                      className={cn(
-                        "w-[80vw] max-w-none object-cover lg:h-full lg:w-full",
-                        idx === 0 &&
-                          "[&_img]:[view-transition-name:image-expand]",
-                      )}
-                    />
-                  </div>
-                  {shouldShowButton && (
-                    <ZoomButton
-                      className={clsx(
-                        "absolute top-2 right-2 md:top-4 md:right-4",
-                        zoomButtonVisibility === "hover" &&
-                          "opacity-0 group-hover:opacity-100",
-                      )}
-                      onClick={() => {
-                        setZoomMediaId(med.id);
-                        setZoomModalOpen(true);
-                      }}
-                    />
-                  )}
-                </div>
-              );
-            })}
-          </div>
-          {shouldLimitMedia && !expanded && (
-            <button
-              type="button"
-              className="absolute right-0 bottom-0 left-0 flex cursor-pointer items-end justify-center bg-linear-to-t from-white/80 via-white/60 to-transparent pt-50 pb-10 font-medium text-body transition-opacity"
-              onClick={() => setExpanded(true)}
-              aria-label={`${showMoreText} (+${hiddenCount})`}
-            >
-              <span className="flex flex-col items-center gap-1">
-                <CaretDownIcon className="h-4 w-4" />
-                <span className="underline underline-offset-4">
-                  {`${showMoreText} (+${hiddenCount})`}
-                </span>
-              </span>
-            </button>
-          )}
-          {shouldLimitMedia && expanded && (
-            <button
-              type="button"
-              className="mt-6 flex w-full cursor-pointer items-center justify-center gap-1 py-2 font-medium lg:mt-2"
-              onClick={() => setExpanded(false)}
-              aria-label={showLessText}
-            >
-              <span className="flex flex-col items-center gap-1">
-                <CaretUpIcon className="h-4 w-4" />
-                <span className="underline underline-offset-4">
-                  {showLessText}
-                </span>
-              </span>
-            </button>
-          )}
-        </div>
-        {enableZoom && (
-          <ZoomModal
-            media={displayMedia}
-            zoomMediaId={zoomMediaId}
-            setZoomMediaId={setZoomMediaId}
-            open={zoomModalOpen}
-            onOpenChange={setZoomModalOpen}
-          />
-        )}
-      </>
-    );
-  }
-
-  return (
-    <div className="product-media-slider overflow-hidden">
-      <div
-        className={clsx(
-          "flex items-start gap-4 overflow-hidden [--thumbs-width:0px]",
-          showThumbnails && "md:[--thumbs-width:8rem]",
-        )}
-      >
-        {showThumbnails && (
-          <div
-            className={clsx(
-              "hidden shrink-0 md:block",
-              "h-112.5 w-[calc(var(--thumbs-width,0px)-1rem)]",
-              "opacity-0 transition-opacity duration-300",
-            )}
-          >
-            <Swiper
-              onSwiper={setThumbsSwiper}
-              direction="vertical"
-              spaceBetween={8}
-              slidesPerView={5}
-              watchSlidesProgress
-              rewind
-              freeMode
-              className="h-full w-full overflow-visible"
-              onInit={(sw) => {
-                sw.el.parentElement.style.opacity = "1";
-              }}
-              modules={[Navigation, Thumbs, FreeMode]}
-            >
-              {displayMedia.map(
-                ({ id, previewImage, alt, mediaContentType }) => {
-                  return (
-                    <SwiperSlide
-                      key={id}
-                      className={cn(
-                        "relative",
-                        "h-auto! cursor-pointer border border-transparent p-1 transition-colors",
-                        "[&.swiper-slide-thumb-active]:border-line",
-                      )}
-                    >
-                      <Image
-                        data={{
-                          ...previewImage,
-                          altText: alt || "Product image",
-                        }}
-                        loading="lazy"
-                        width={200}
-                        aspectRatio="1/1"
-                        className="h-auto w-full object-cover"
-                        sizes="auto"
-                      />
-                      {mediaContentType === "VIDEO" && (
-                        <div className="absolute right-2 bottom-2 bg-gray-900 p-0.5 text-white">
-                          <VideoCameraIcon className="h-4 w-4" />
-                        </div>
-                      )}
-                    </SwiperSlide>
-                  );
-                },
-              )}
-            </Swiper>
-          </div>
-        )}
-        <div className="relative w-[calc(100%-var(--thumbs-width,0px))]">
-          <Swiper
-            onSwiper={setSwiper}
-            thumbs={{ swiper: thumbsSwiper }}
-            slidesPerView={1}
-            spaceBetween={4}
-            autoHeight
-            loop
-            navigation={{
-              nextEl: ".media_slider__next",
-              prevEl: ".media_slider__prev",
-            }}
-            pagination={{ type: "fraction" }}
-            modules={[Pagination, Navigation, Thumbs]}
-            className="overflow-visible pb-10 md:overflow-hidden md:pb-0 md:[&_.swiper-pagination]:hidden"
-          >
-            {displayMedia.map((med, idx) => {
-              return (
-                <SwiperSlide key={med.id} className="group bg-gray-100">
-                  <div
-                    onClick={
-                      canClickImage
-                        ? () => {
-                            setZoomMediaId(med.id);
-                            setZoomModalOpen(true);
-                          }
-                        : undefined
-                    }
-                    className={canClickImage ? "cursor-zoom-in" : ""}
-                  >
-                    <Media
-                      media={med}
-                      imageAspectRatio={imageAspectRatio}
-                      index={idx}
-                      className={
-                        idx === 0 &&
-                        "[&_img]:[view-transition-name:image-expand]"
-                      }
-                    />
-                  </div>
-                  {shouldShowButton && (
-                    <ZoomButton
-                      className={clsx(
-                        "absolute top-2 right-2 md:top-6 md:right-6",
-                        zoomButtonVisibility === "hover" &&
-                          "opacity-0 group-hover:opacity-100",
-                      )}
-                      onClick={() => {
-                        setZoomMediaId(med.id);
-                        setZoomModalOpen(true);
-                      }}
-                    />
-                  )}
-                </SwiperSlide>
-              );
-            })}
-          </Swiper>
-          <div className="absolute right-6 bottom-6 z-1 hidden items-center gap-2 md:flex">
-            <button
-              type="button"
-              className="media_slider__prev left-6 border border-transparent bg-white p-2 text-center text-gray-900 transition-all duration-200 hover:bg-gray-800 hover:text-white disabled:cursor-not-allowed disabled:text-body-subtle"
-            >
-              <ArrowLeftIcon className="h-4.5 w-4.5" />
-            </button>
-            <button
-              type="button"
-              className="media_slider__next right-6 border border-transparent bg-white p-2 text-center text-gray-900 transition-all duration-200 hover:bg-gray-800 hover:text-white disabled:cursor-not-allowed disabled:text-body-subtle"
-            >
-              <ArrowRightIcon className="h-4.5 w-4.5" />
-            </button>
-          </div>
-        </div>
-      </div>
-      {enableZoom && (
-        <ZoomModal
-          media={displayMedia}
-          zoomMediaId={zoomMediaId}
-          setZoomMediaId={setZoomMediaId}
-          open={zoomModalOpen}
-          onOpenChange={setZoomModalOpen}
-        />
-      )}
-    </div>
-  );
-}
-
-function Media({
-  media,
-  imageAspectRatio,
-  index,
-  className,
-}: {
-  media: MediaFragment;
-  imageAspectRatio: ImageAspectRatio;
-  index: number;
-  className?: string;
-}) {
-  if (media.mediaContentType === "IMAGE") {
-    const { image, alt } = media as Media_MediaImage_Fragment;
-    return (
-      <Image
-        data={{ ...image, altText: alt || "Product image" }}
-        loading={index === 0 ? "eager" : "lazy"}
-        className={cn("h-auto w-full object-cover", className)}
-        width={2048}
-        aspectRatio={calculateAspectRatio(image, imageAspectRatio)}
-        sizes="auto"
-      />
-    );
-  }
-  if (media.mediaContentType === "VIDEO") {
-    const mediaVideo = media as Media_Video_Fragment;
-    return (
-      <video
-        controls
-        aria-label={mediaVideo.alt || "Product video"}
-        className={cn("h-auto lg:aspect-video w-full object-cover", className)}
-        style={{ aspectRatio: imageAspectRatio }}
-        onError={console.error}
-      >
-        <track
-          kind="captions"
-          src={mediaVideo.sources[0].url}
-          label="English"
-          srcLang="en"
-          default
-        />
-        <source src={mediaVideo.sources[0].url} type="video/mp4" />
-      </video>
-    );
-  }
-  return null;
-}
-
-function getSelectedVariantMediaIndex(
-  media: MediaFragment[],
-  selectedVariant: ProductVariantFragment,
-) {
-  if (!selectedVariant) {
-    return 0;
-  }
-  const mediaUrl = selectedVariant.image?.url;
-  return media.findIndex((med) => med.previewImage?.url === mediaUrl);
-}
