+++ app/components/product/product-media/media-grid.tsx
@@ -1,0 +1,155 @@
+import { CaretDownIcon, CaretUpIcon } from "@phosphor-icons/react";
+import clsx from "clsx";
+import { useState } from "react";
+import type { MediaFragment } from "storefront-api.generated";
+import type { ImageAspectRatio } from "~/types/others";
+import { cn } from "~/utils/cn";
+import { MediaItem } from "./media-item";
+import { ZoomButton, ZoomModal } from "./media-zoom";
+import { mediaGridVariants } from "./utils";
+
+interface MediaGridProps {
+  displayMedia: MediaFragment[];
+  gridSize: "1x1" | "2x2" | "mix";
+  imageAspectRatio?: ImageAspectRatio;
+  enableZoom?: boolean;
+  zoomTrigger?: "image" | "button" | "both";
+  zoomButtonVisibility?: "always" | "hover";
+  initialMediaCount?: number;
+  showMoreText?: string;
+  showLessText?: string;
+  expanded: boolean;
+  setExpanded: (expanded: boolean) => void;
+}
+
+export function MediaGrid({
+  displayMedia,
+  gridSize,
+  imageAspectRatio,
+  enableZoom,
+  zoomTrigger = "button",
+  zoomButtonVisibility = "hover",
+  initialMediaCount = 0,
+  showMoreText = "Show more",
+  showLessText = "Show less",
+  expanded,
+  setExpanded,
+}: MediaGridProps) {
+  const [zoomMediaId, setZoomMediaId] = useState<string | null>(null);
+  const [zoomModalOpen, setZoomModalOpen] = useState(false);
+
+  const shouldShowButton =
+    enableZoom && (zoomTrigger === "button" || zoomTrigger === "both");
+  const canClickImage =
+    enableZoom && (zoomTrigger === "image" || zoomTrigger === "both");
+
+  const shouldLimitMedia =
+    initialMediaCount > 0 && displayMedia.length > initialMediaCount;
+  const visibleMedia =
+    shouldLimitMedia && !expanded
+      ? displayMedia.slice(0, initialMediaCount)
+      : displayMedia;
+  const hiddenCount = displayMedia.length - visibleMedia.length;
+
+  return (
+    <>
+      <div className="relative">
+        <div className={mediaGridVariants({ gridSize })}>
+          {visibleMedia.map((med, idx) => {
+            const isLast = idx === visibleMedia.length - 1;
+            return (
+              <div
+                key={med.id}
+                className={clsx(
+                  "group relative",
+                  gridSize === "2x2" &&
+                    isLast &&
+                    visibleMedia.length % 2 === 1 &&
+                    "2xl:col-span-2",
+                  gridSize === "mix" &&
+                    (idx % 3 === 0 || (isLast && idx % 3 === 1)) &&
+                    "lg:col-span-2",
+                )}
+              >
+                <div
+                  onClick={
+                    canClickImage
+                      ? () => {
+                          setZoomMediaId(med.id);
+                          setZoomModalOpen(true);
+                        }
+                      : undefined
+                  }
+                  className={canClickImage ? "cursor-zoom-in" : ""}
+                >
+                  <MediaItem
+                    media={med}
+                    imageAspectRatio={imageAspectRatio}
+                    index={idx}
+                    className={cn(
+                      "w-[80vw] max-w-none object-cover lg:h-full lg:w-full",
+                      idx === 0 &&
+                        "[&_img]:[view-transition-name:image-expand]",
+                    )}
+                  />
+                </div>
+                {shouldShowButton && (
+                  <ZoomButton
+                    className={clsx(
+                      "absolute top-2 right-2 md:top-4 md:right-4",
+                      zoomButtonVisibility === "hover" &&
+                        "opacity-0 group-hover:opacity-100",
+                    )}
+                    onClick={() => {
+                      setZoomMediaId(med.id);
+                      setZoomModalOpen(true);
+                    }}
+                  />
+                )}
+              </div>
+            );
+          })}
+        </div>
+        {shouldLimitMedia && !expanded && (
+          <button
+            type="button"
+            className="absolute right-0 bottom-0 left-0 flex cursor-pointer items-end justify-center bg-linear-to-t from-white/80 via-white/60 to-transparent pt-50 pb-10 font-medium text-body transition-opacity"
+            onClick={() => setExpanded(true)}
+            aria-label={`${showMoreText} (+${hiddenCount})`}
+          >
+            <span className="flex flex-col items-center gap-1">
+              <CaretDownIcon className="h-4 w-4" />
+              <span className="underline underline-offset-4">
+                {`${showMoreText} (+${hiddenCount})`}
+              </span>
+            </span>
+          </button>
+        )}
+        {shouldLimitMedia && expanded && (
+          <button
+            type="button"
+            className="mt-6 flex w-full cursor-pointer items-center justify-center gap-1 py-2 font-medium lg:mt-2"
+            onClick={() => setExpanded(false)}
+            aria-label={showLessText}
+          >
+            <span className="flex flex-col items-center gap-1">
+              <CaretUpIcon className="h-4 w-4" />
+              <span className="underline underline-offset-4">
+                {showLessText}
+              </span>
+            </span>
+          </button>
+        )}
+      </div>
+      {enableZoom && (
+        <ZoomModal
+          media={displayMedia}
+          zoomMediaId={zoomMediaId}
+          setZoomMediaId={setZoomMediaId}
+          open={zoomModalOpen}
+          onOpenChange={setZoomModalOpen}
+        />
+      )}
+    </>
+  );
+}
